// frontend/Angular-workflow-diagram2-main/src/app/api/model/crearSolicitudRequest.ts
/**
 * Workflow Departamental API
 *
 * Contact: soporte@workflow.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface CrearSolicitudRequest { 
    titulo: string;
    descripcion: string;
    prioridad: CrearSolicitudRequest.PrioridadEnum;
    departamentoDestino: string;
    workflowDefinitionId?: string;
}
export namespace CrearSolicitudRequest {
    export const PrioridadEnum = {
        Baja: 'BAJA',
        Media: 'MEDIA',
        Alta: 'ALTA',
        Urgente: 'URGENTE'
    } as const;
    export type PrioridadEnum = typeof PrioridadEnum[keyof typeof PrioridadEnum];
}